implement feedback on UI
