+++ src/components/PaymentSection/TableHeader.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export const TableHeader = () => (
+  <div className="table_header flex">
+    <span className="col col1">Item Type</span>
+    <span className="col col2">Price</span>
+    <span className="col col3">Transaction No</span>
+    <span className="col col4">Time</span>
+    <span className="col col5">Status</span>
+    <span className="col col6">Status</span>
+  </div>
+);
