+++ src/styles/globalstyles.js
@@ -1,9 +1,20 @@
 import styled, { createGlobalStyle } from 'styled-components';
 
 export const GlobalStyle = createGlobalStyle`
+  @font-face {
+    font-family: SegoeUI;
+    src:
+      local("Segoe UI"),
+      url(//c.s-microsoft.com/static/fonts/segoe-ui/west-european/normal/latest.woff2) format("woff2"),
+      url(//c.s-microsoft.com/static/fonts/segoe-ui/west-european/normal/latest.woff) format("woff"),
+      url(//c.s-microsoft.com/static/fonts/segoe-ui/west-european/normal/latest.ttf) format("truetype");
+    font-weight: 400;
+  }
+
   html {
     -webkit-box-sizing: border-box;
             box-sizing: border-box;
+       scroll-behavior: smooth;
   }
 
   *, *:before, *:after {
@@ -84,7 +95,7 @@ export const ContainerWrapper = styled.section`
 
   display: flex;
   justify-content: start;
-  height: 100%
+  height: 100%;
 
   .content_wrapper {
     flex-grow: 1;
